requirements.txt
pygame==1.9.4
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
